### 2022-06-29 #4525 @editorialbot
[PRE REVIEW]: dtrackr: An R package for tracking the provenance of data
**Submitting author:** @robchallen (<a href="http://orcid.org/0000-0002-5504-7768">Robert Challen</a>)
**Repository:** https://github.com/terminological/dtrackr
**Branch with paper.md** (empty if default branch): release-0.2.4
**Version:** 0.2.4
