### 2022-06-29 a48fa6b
non functional metadata fixes for CRAN. staying with version 0.2.3
