+++ vignettes/dtrackr.bib
@@ -58,7 +58,7 @@
   author = {Dua, Dheeru and Graff, Casey},
   date = {2017},
   publisher = {{University of California, Irvine, School of Information and Computer Sciences}},
-  url = {http://archive.ics.uci.edu/ml}
+  url = {http://archive.ics.uci.edu/ml/}
 }
 
 @article{benchimolREportingStudiesConducted2015,
