+++ cran-comments.md
@@ -6,11 +6,22 @@ Github actions environments
 * ubuntu-20-04 with R 4.2.0
 * macOS-11 with R 4.1.0
 * windows-2022 with R 4.1.0
+* ubuntu-18.04 with R r-devel
 
 ## R CMD check results
 There were no ERRORs or WARNINGs or NOTEs. 
+(see https://github.com/terminological/dtrackr/actions for details)
 
 ## Downstream dependencies
 There are currently no downstream dependencies for this package.
 
-This is a new submission
+## Other info
+
+The project is hosted on github - https://github.com/terminological/dtrackr
+
+This is a re-submission of an as-yet unpublished package correcting the following issues:
+* Please only ship the CRAN template for the MIT license.
+* Please single quote software names in the Description field.
+* Please change http --> https, add trailing slashes, or follow moved content as appropriate.
+* The title field should be in title case. 
+* The Description field should not start with the package name, 'This package' or similar. Please fix and resubmit.
