### 2022-06-29 21ee615
non functional metadata fixes for CRAN. staying with version 0.2.3
