+++ .github/workflows/R-CMD-check.yaml
@@ -48,7 +48,7 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@v2
 
-      - uses: r-lib/actions/setup-r-dependencies@v1
+      - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 3
           extra-packages: rcmdcheck
