### 2022-07-01 #4525 @Kevin-Mattheus-Moerman
@robchallen I have just assigned the `query scope` label which means I have triggered a scope review by our editorial board. This is typical for submission around this size. We will review if this work complies with our ["substantial scholarly effort" criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), and if it has sufficient functionality and enough of a research application. This additional scope review will take about 2 weeks to complete.
