### 2022-07-03 6d70c6d
cran fixes second round. version bump to 0.2.4.
