+++ R/dot.R
@@ -1,14 +1,14 @@
 
-#' Convert dot to svg
+#' Convert Graphviz dot content to a SVG
 #'
-#' Convert a digraph as string to SVG as string
+#' Convert a graphviz dot digraph as string to SVG as string
 #'
 #' @param dot - a graphviz dot string
 #'
 #' @return the SVG as a string
 #' @export
 #'
-#' @examples dot2svg("digraph {A->B}")
+#' @examples dot2svg("digraph { A->B }")
 dot2svg <- function(dot) {
 
   if(!requireNamespace("V8", quietly = TRUE)) stop("V8 package is required", call. = FALSE)
@@ -41,6 +41,8 @@ dot2svg <- function(dot) {
 
 #' Standard paper sizes
 #'
+#' A list of standard paper sizes
+#'
 #' @export
 std_size = list(
   A4 = list(width=8.25,height=11.75,rot=0),
@@ -62,62 +64,23 @@ std_size = list(
 landscape = function(size) {return(list(width=size$height, height = size$width, rot=270))}
 
 
-.relPath = function(pathname) {
-  # Split the two pathnames into their components
-  relativeTo <- unlist(strsplit(getwd(), split="[\\/]"))
-  pathname <- unlist(strsplit(pathname, split="[\\/]"))
-  pathnameC <- pathname
-
-  # 1. Check that the pathnames are "compatible".
-  if (!identical(relativeTo[1L], pathnameC[1L])) {
-    pathname <- paste(pathname, collapse="/")
-
-    return(pathname)
-  }
-
-  # 2. Remove all matching components in 'relativeTo' and 'pathname'.
-  #    The removed parts constitute their common path.
-  for (kk in seq_along(relativeTo)) {
-    aPart <- relativeTo[1]
-    bPart <- pathnameC[1]
-    if (!identical(aPart, bPart))
-      break
-
-    relativeTo <- relativeTo[-1L]
-    pathname <- pathname[-1L]
-    pathnameC <- pathnameC[-1L]
-  }
-
-  # 3. If there are more components in 'relativeTo', this means that the
-  #    rest of 'relativeTo' is in a different subdirectory than 'pathname'.
-  prefix <- rep("..", length.out=length(relativeTo))
-
-  pathname <- c(prefix, pathname)
-  pathname <- paste(pathname, collapse="/")
-
-  if (pathname == "")
-    pathname <- "."
-
-  pathname
-}
-
-#' Save DOT to a file
+#' Save DOT content to a file
 #'
-#' Convert a digraph and save as a file
+#' Convert a digraph in dot file to SVG and save it to an output file
 #'
 #' @param dot - a graphviz dot string
-#' @param filename - the root of the desired filename (minus extension)
+#' @param filename - the full path of the filename (minus extension for multiple formats)
 #' @param size - a list of length and width in inches e.g. a std_size
 #' @param maxWidth - a width in inches is size is not defined
 #' @param maxHeight - a height in inches if size is not defined
 #' @param rot - an angle of rotation for the saved file if size is not defined
 #' @param formats - some of "pdf","dot","svg","png","ps"
 #'
-#' @return the SVG as a string
+#' @return a list with items `paths` with the absolute paths of the saved files, and svg as the SVG string of the rendered dot file.
 #' @export
 #'
-#' @examples dot2svg("digraph {A->B}")
-save_dot = function(dot,filename,size = std_size$half, maxWidth = size$width, maxHeight = size$height, rot=size$rot, formats=c("dot","png","pdf","svg")) {
+#' @examples dot2svg("digraph {A->B} ")
[…]
