+++ R/dtrackr.R
@@ -196,14 +196,67 @@
   return(out)
 }
 
+#' Print a history graph to the console
+#'
+#' @param x a dtrackr history graph (e.g. output from [p_get()])
+#' @param ... not used
+#'
+#' @return nothing
+#' @export
+#'
+#' @examples
+#' library(dplyr)
+#' iris %>% comment("hello {.total} rows") %>% history() %>% print()
+print.trackr_graph = function(x, ...) {
+  graph = x
+  ranks = length(unique(graph$nodes$.rank))
+  tags = paste0(unique(graph$tags$.tag),collapse="; ")
+  if (tags=="") tags="<none>"
+  if (is.null(graph$excluded)) {
+    excluded = "<not capturing exclusions>"
+  } else {
+    excluded = sum(sapply(graph$excluded$.excluded, nrow ), na.rm = TRUE)
+  }
+
+  tmp = c(
+    "dtrackr history:",
+    glue::glue("number of flowchart steps: {ranks} (approx)"),
+    glue::glue("tags defined: {tags}"),
+    glue::glue("items excluded so far: {excluded}")
+  )
+  if (isTRUE(graph$paused)) {
+   tmp = c(tmp,"TRACKING IS PAUSED")
+  }
+  cat(tmp,sep = "\n")
+}
+
+#' Plots a history graph as html
+#'
+#' @param x a dtrackr history graph (e.g. output from [history()])
+#' @param fill - the default node fill colour
+#' @param fontsize - the default font size
+#' @param colour - the default font colour
+#' @param ... not used
+#'
+#' @return HTML displayed
+#' @export
+#'
+#' @examples
+#' library(dplyr)
+#' iris %>% comment("hello {.total} rows") %>% history() %>% plot()
+plot.trackr_graph = function(x, fill="lightgrey", fontsize="8", colour="black", ...) {
+  graph = x
+  outgraph = .graph2dot(graph, fill, fontsize, colour, ...)
+  htmltools::HTML(dot2svg(outgraph)) %>% htmltools::html_print()
+}
+
 ## Basic operations ----
 
 .emptyGraph = function(exclusions = FALSE,...) {
   nodes=tibble::tibble(.id=integer(),.rank=integer(),.strata=character(),.label=character(),.type=character())
   edges=tibble::tibble(.to=integer(),.from=integer(),.rel=character(), .strata=character()) # empty edges
   head=tibble::tibble(.from=integer(), .strata=character())
-  return(
-    list(
+  tmp = list(
       nodes=nodes,
       edges=edges,
       head=head,
@@ -211,7 +264,9 @@
       paused=FALSE,
       tags=tibble::tibble(.tag=character(), .content=list())
       # excluded is set in p_capture_exclusions and will not be present all the time
-  ))
+  )
+  class(tmp) = unique(c("trackr_graph",class(tmp)))
+  return(tmp)
 }
 
 .untrack = function(.data) {
@@ -301,15 +356,16 @@
 #'
 #' @param .data - a dataframe which may be grouped
 #' @param .messages - a character vector of glue specifications. A glue specification can refer to any grouping variables of .data, or any variables defined in the calling environment, the \{.total\} variable which is the count of all rows,
-#' the \{.count\} variable which is the count of rows in the current group and the \{.strata\} which describes the current group. Defaults to the value of getOption("dtrackr.default_message").
+#' the \{.count\} variable which is the count of rows in the current group and the \{.strata\} which describes the current group. Defaults to the value of `getOption("dtrackr.default_message")`.
[…]
