### 2022-07-04 ee0a6d9
3rd round CRAN fixes
