### 2022-07-07 #4525 @robchallen
@Kevin-Mattheus-Moerman just a quick note. I've been through the (somewhat tortuous) process of getting this library on CRAN. This required a minor rework of some of the functions, and documentation fixes, so I've bumped the version to 0.2.4 (which has a branch and release).

I don't know if that influences the additional scope review or not, bout thought you should be aware anyway.
