### 2022-07-30 #4525 @robchallen
Yes.

It is on CRAN now (with the statement that it is in early development and
not yet on cran, sigh...).

I'll put a bug fix release through cran over the next couple of weeks to
fix this and anything else that pops up.

On Fri, 29 Jul 2022, 22:41 Kevin Mattheus Moerman, ***@***.***>
wrote:

> @robchallen <https://github.com/robchallen> the board has decided this
> work may be in scope for JOSS. Hence I've removed the quary-scope label.
> The board did have some recommendations, i.e. that the work is submitted to
> CRAN (which it sounds like you've done). If so, you can probably remove the
> statement dtrackr is in early development. It is not yet available on
> CRAN. from the README.
>
> We will now proceed to try to find a handling editor for this submission.
> Unfortunately we have a backlog of submissions at the moment and our
> editors are handling many other submissions. Therefore we will waitlist
> this work for the moment until a suitable editor becomes available.
>
