### 2022-08-24 #4707 @editorialbot
[REVIEW]: dtrackr: An R package for tracking the provenance of data
**Submitting author:** @robchallen (<a href="http://orcid.org/0000-0002-5504-7768">Robert Challen</a>)
**Repository:** https://github.com/terminological/dtrackr
**Branch with paper.md** (empty if default branch): main
**Version:** 0.2.5-joss
