### 2022-08-25 #4707 @robchallen
I just wanted to mention that in the branch here under review,  the README.md instructions say that "dtrackr" is not yet on CRAN.

This is not correct and the version in the joss review branch (release-0.2.4) of dtrackr is now on CRAN and can be installed with a install.packages("dtrackr") command.

I'm holding back on updating the documentation until I get comments back from you guys. Then I'll make any outstanding fixes update the documentation and push a 0.2.5 release through the CRAN submission process (with the updated installation instructions showing it is on CRAN).

I've just made the exact same mistake on another package I'm submitting to CRAN :-)
