### 2022-09-22 terminological/dtrackr#10 @craig-willis
JOSS review - License
Review issue: https://github.com/openjournals/joss-reviews/issues/4707
Branch reviewed: release-0.2.4

**Problem**
JOSS requires that your repository include the full-text of an OSI approved license. See the [Software License](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license) section of the review criteria. The current [LICENSE](https://github.com/terminological/dtrackr/blob/release-0.2.4/LICENSE) file does not appear to be an OSI approved license.  The [DESCRIPTION](https://github.com/terminological/dtrackr/blob/b79ee31fe4effd2c40c2d634a237f16fc0440caf/DESCRIPTION#L14) indicates an MIT license.

If the MIT license has been selected, consider updating the LICENSE file based on the [OSI template](https://opensource.org/licenses/MIT). (Also, I'm not sure what "MIT + file LICENSE" means in the DESCRIPTION).
