### 2022-09-22 terminological/dtrackr#11 @craig-willis
JOSS review - Community guidelines
Review issue: https://github.com/openjournals/joss-reviews/issues/4707
Branch reviewed: release-0.2.4

**Problem**
JOSS requires clear guidelines for third-parties wishing to contribute to the software.  See [Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) for more information.

A common approach is to add a [CONTRIBUTING](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) file to the repository.
