### 2022-09-22 terminological/dtrackr#11 @craig-willis
Just noticed the `.github/CONTRIBUTING.md` and that it appears on the package website. I'm accustomed to seeing this in the repo root, but expect this meets the JOSS requirement.
