### 2022-09-22 terminological/dtrackr#12 @craig-willis
JOSS review - Installation instructions
Review issue: https://github.com/openjournals/joss-reviews/issues/4707
Branch reviewed: release-0.2.4

**Problem**:
The JOSS guidelines require clear [Installation instructions](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions).  I attempted to install `dtrackr` on a fresh Ubuntu 20.04 (via Docker) with R 4.2.1 and encountered several errors due to missing OS packages. These could be inferred from the required R packages, but it would be helpful to have a complete list in the installation instructions. Additional packages and OS dependencies were required to run the automated tests and could also be clearer.

What I had to install:
* R (of course)
* librsvg2-bin librsvg2-dev libv8-dev libcurl4-openssl-dev(via apt)

For automated tests:
* libxml2 (via apt)
* tidyverse

Consider adding a comprehensive list of dependencies as well as supported OS and R versions to the "Installing" section of the readme.
