### 2022-09-22 terminological/dtrackr#13 @craig-willis
JOSS review - Automated tests
Review issue: https://github.com/openjournals/joss-reviews/issues/4707
Branch reviewed: release-0.2.4

**Problem**
JOSS strongly recommends using [Automated tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests), which this repo supports through the use of `testthat` and Github actions. I expect that the current implementation meets this requirement, but do have some comments:
* The completeness of the tests is unclear with the current implementation.
* Github action [output](https://github.com/terminological/dtrackr/actions/runs/2584445776/jobs/3988803240) suggests a potential problem with package conflicts.
* Automated tests are not documented (outside of the Github action configuration).

**Test Coverage**
I'm not very familiar with R test automation, but I was able to combine `testthat` with `covr` to get a brief coverage report:
```
covr::package_coverage(
  type = "none",
  code = "testthat::test_check(
'dtrackr'
)")

dtrackr Coverage: 56.18%
R/dot.R: 13.51%
R/dtrackr.R: 60.72%
```
Without looking at the test implementations in detail, this suggests that not all of the `dtrackr` functions are currently tested. JOSS makes no recommendations for automated test coverage rates.

**Package conflicts**
The test output shows conflicts between `dplyr` and `dtrackr` functions. I may be misunderstanding, but shouldn't `dtrackr` be load after `tidyverse` to test the overridden `dtrackr` methods?

**Documentation**
It took a little effort to figure out how to run the automated tests. For a contributor or user, it might be helpful to have a section in the documentation briefly describing the test automation approach (testthat), Github action, as well as how to install and run locally (see my comment about dependencies in https://github.com/terminological/dtrackr/issues/12).
