### 2022-09-22 terminological/dtrackr#14 @craig-willis
JOSS review - Functionality documentation
Review issue: https://github.com/openjournals/joss-reviews/issues/4707
Branch reviewed: release-0.2.4

**Problem**
It's not immediately clear to me from the documentation which `dplyr` and `tidyr` operations are supported by `dtrackr` or why some are missing. This can be determined only by comparing the list of `p_*` functions and inspecting the source code comments.

I'm not a regular user of `dplyr`, but based on comparison to https://dplyr.tidyverse.org/reference/ see the following operations are not supported by `dtrackr`: `count`, `tally`, `glimpse`, `group*`, `pull`, `summarize`, `slice*`, `bind_cols`, `union_all`, `nest_join`, `rowwise`.

Code comments indicate `count / tally - not included as summarisation step, not generally part of a dplyr pipeline` and `group_map/group_walk -- not included as does not return DF`.

The `pivot_wider` and `pivot_longer` appear to be the only operations supported from `tidyr`.

As a user, it would be helpful to have a quick reference of supported operations and explanations of why certain operations are/are not included. This might also be an opportunity to highlight the relationship between `dplyr::filter` and `exclude_all`, since it's not strictly wrapping the `dplyr` operation.
