### 2022-09-22 terminological/dtrackr#15 @craig-willis
JOSS review - Software paper
Review issue: https://github.com/openjournals/joss-reviews/issues/4707

Below is my feedback on the "Software paper" section of the JOSS review checklist:

**Summary**
I question whether `dtrackr` can be described as a wrapper around the `tidyverse` collection of packages and think it may better be described as wrapper around `dplyr`.  There may be other opportunities to support data provenance in tidyverse that cannot be done by instrumenting dplyr functions (e.g., read files).

**State of the field**
`dtrackr` is positioned as a provenance tool yet the paper includes no references to existing work in the area of computational provenance, transparency, and reproducibility. While I believe `dtrackr` functionality to be unique, the tool is not positioned relative to other packages in the R community or beyond.  I'd suggest looking at the following articles and considering how `dtrackr` relates to other efforts.

* Pimentel et al. 2019. A Survey on Collecting, Managing, and Analyzing Provenance from Scripts. ACM Comput. Surv. 52, 3, Article 47 (May 2020), 38 pages. https://doi-org/10.1145/3311955
* Alter et a. 2021. Capturing Data Provenance from Statistical Software. International Journal of Digital Curation. https://doi.org/10.2218/ijdc.v16i1.763.

The absence of comparisons to other provenance tools to some degree raises a question about the scholarly aspects of this work (i.e., is this just a utility or intended to contribute to broader work in computational provenance, transparency, reproducibility). I can imagine how a widely-used provenance-aware dyplr might fit into this bigger picture.
