### 2022-09-22 terminological/dtrackr#10 @robchallen
Many thanks for this and other comments which are all very helpful.

I went through a few iterations with the LICENSE files. CRAN have some specific ways they want the files which I now sort of understand, but required me taking out the original. I've added in a LICENCE.md file into this repo for github, which is ignored for CRAN submission. The 2 line LICENSE text file and DESCRIPTION entry "MIT + file LICENSE" is the correct incantation for CRAN.
