### 2022-09-22 terminological/dtrackr#11 @robchallen
Thanks. I understand re the comments being too personal. I have revised the CONTRIBUTING section based on a template I found below.

This Code of Conduct is adapted from the [Contributor Covenant][homepage], version 1.4, available at [http://contributor-covenant.org/version/1/4][version]
