### 2022-09-23 terminological/dtrackr#10 @craig-willis
Ah, that makes sense now. I wasn't aware of the CRAN license requirements but see now that this is a common approach. Did you push your change to Github?  I don't see a new commit.
