### 2022-09-23 terminological/dtrackr#10 @robchallen
Not yet.
I'll try and bundle a few things together, so it doesn't come at you piecemeal.
It will probably will take me a few days.
Do you close these issues or do I?
