### 2022-09-23 terminological/dtrackr#10 @craig-willis
Sounds good.

As for closing the issues, I've confirmed with the editor that either of is can close an issue after it's been addressed. I'm happy to review any changes (or explanation of why changes don't need to be made) and close them. You're also welcome to close them and I'll let you know if I have any further concerns.

For JOSS, the main thing is that I communicate any concerns -- particularly that might block acceptance -- via the primary review issue. I don't plan on putting anything there until you've had a chance to respond to the open issues in this repo.
