### 2022-09-30 terminological/dtrackr#12 @robchallen
Thank you. I've rewritten quite a lot of the README.Rmd file to address both this and to make the style of the README more formal.

Maybe you are aware of if already but I found `pak::pkg_system_requirements()` to be useful for this and I have signposted people to it for more unusual Linux distributions.

I've added a section on developing and running the automated tests
