+++ tests/testthat/test-p_comment.R
@@ -1,4 +1,6 @@
 library(tidyverse)
+library(dtrackr)
+
 graphContains = function(g, .strata, .message) {
   return(g$nodes %>% filter(.strata==.strata,.label %>% str_detect(paste0("(^|>)",fixed(.message),"<"))) %>% nrow() > 0)
 }
