+++ R/dot.R
@@ -8,7 +8,8 @@
 #' @return the SVG as a string
 #' @export
 #'
-#' @examples dot2svg("digraph { A->B }")
+#' @examples
+#' dot2svg("digraph { A->B }")
 dot2svg <- function(dot) {
 
   if(!requireNamespace("V8", quietly = TRUE)) stop("V8 package is required", call. = FALSE)
@@ -73,14 +74,15 @@ landscape = function(size) {return(list(width=size$height, height = size$width, 
 #' @param size - a list of length and width in inches e.g. a std_size
 #' @param maxWidth - a width in inches is size is not defined
 #' @param maxHeight - a height in inches if size is not defined
-#' @param rot - an angle of rotation for the saved file if size is not defined
 #' @param formats - some of "pdf","dot","svg","png","ps"
+#' @param ... ignored
 #'
-#' @return a list with items `paths` with the absolute paths of the saved files, and svg as the SVG string of the rendered dot file.
+#' @return a list with items `paths` with the absolute paths of the saved files as a named list, and `svg` as the SVG string of the rendered dot file.
 #' @export
 #'
-#' @examples dot2svg("digraph {A->B} ")
-save_dot = function(dot, filename, size = std_size$half, maxWidth = size$width, maxHeight = size$height, rot=size$rot, formats=c("dot","png","pdf","svg")) {
+#' @examples
+#' save_dot("digraph {A->B}",tempfile())
+save_dot = function(dot, filename, size = std_size$half, maxWidth = size$width, maxHeight = size$height, formats=c("dot","png","pdf","svg"), ...) {
 
   tmp = filename %>% fs::path_ext()
   if(tmp %in% c("dot","png","pdf","svg","ps")) formats=tmp
@@ -97,63 +99,75 @@ save_dot = function(dot, filename, size = std_size$half, maxWidth = size$width, 
   svg = dot %>% dot2svg()
   defaultWidth = (svg %>% stringr::str_match("width=\"([0-9]*)pt\""))[1,2] %>% as.numeric()
   defaultHeight = (svg %>% stringr::str_match("height=\"([0-9]*)pt\""))[1,2] %>% as.numeric()
-
   aspectRatio = defaultWidth / defaultHeight
+  widthIn = min(maxWidth,maxHeight*aspectRatio,defaultWidth/72)
+  heightIn = min(maxHeight,maxWidth/aspectRatio,defaultHeight/72)
+
+  # resize = function(dpi) {
+  #   svg %>%
+  #     stringr::str_replace_all("width=\"([0-9]*)pt\"", sprintf("width=\"%1.0fpt\"", widthIn*dpi)) %>%
+  #     stringr::str_replace_all("height=\"([0-9]*)pt\"", sprintf("height=\"%1.0fpt\"", heightIn*dpi)) # %>%
+  #     # stringr::str_replace_all("viewBox=\"[^\"]+\"", sprintf("viewBox=\"0.00 0.00 %1.2f %1.2f\"", widthIn*dpi, heightIn*dpi))
+  # }
 
   if ("pdf" %in% formats) {
     svg %>% charToRaw() %>% rsvg::rsvg_pdf(
       file = fname("pdf"),
-      width = min(maxWidth,maxHeight*aspectRatio,defaultWidth)*72,
-      height = min(maxHeight,maxWidth/aspectRatio,defaultHeight)*72
+      width = widthIn*72,
+      height = heightIn*72
     )
     try(
       grDevices::embedFonts(fname("pdf")),
       silent=TRUE
     );
-    if (rot!=0 & isNamespaceLoaded("staplr")) {
-      staplr::rotate_pdf(page_rotation = rot,
-        input_filepath = fname("pdf"),
[…]
