+++ tests/testthat/test-examples.R
@@ -1,0 +1,9 @@
+library(tidyverse)
+library(dtrackr)
+
+# test_that("examples run to completion", {
+#   devtools::run_examples()
+#
+# })
+
+test_examples(path = "../..")
