### 2022-09-30 c9a1b14
minor workflow bug fix
