### 2022-09-30 terminological/dtrackr#13 @robchallen
### Package conflicts:

You are right, that was a mistake, however as the dtrackr methods add S3 methods on top of the `dplyr` functions it actually didn't make any difference (although could have done). Anyway 've fixed this issue.

### Documentation:

Sorry that was an effort to set up. I've added a section in the README about running the tests and the CI setup with github workflows.

### Coverage:

One of the so far little used functions in the `R/dot.R` script was not correctly tested, due to a missing example script. This I have fixed. Usefully this threw up a couple of issues, paritcularly saving content to png files, which I have addressed.

In terms of the other coverage, test_check checks the testthat unit tests only. For 'dtrackr' a lot of the automated testing functionality is in the examples in the man pages which are run when a`R CMD check` is done either manually or by the github workflows. Additionally the vignettes are also a third level of testing, again triggered by a R CMD check. This is better documented in the README file

When the 3 types of test are combined together the coverage is much improved:
```R
covr::package_coverage(type="all")
# dtrackr Coverage: 79.50%
# R/dot.R: 78.12%
# R/dtrackr.R: 79.63%

✔ |  51       | examples [17.6 s]
✔ |   2       | group_by [1.7 s]
✔ |  11       | p_comment [2.4 s]
✔ |   7       | p_exclude [2.3 s]
✔ |   5       | p_group_modify [1.4 s]
✔ |   5       | p_include [2.2 s]
✔ |   4       | p_others [1.5 s]
✔ |   5       | p_status [2.1 s]

══ Results ═════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Duration: 31.3 s

[ FAIL 0 | WARN 0 | SKIP 0 | PASS 90 ]

WHOA - stupendous code.
```
Some of the functions that are not tested are very thin wrappers around the `dplyr` functions - essentially simple pass through code.

I'll work towards integrating the `covr` output into the package in the future. I believe there is a badge for this but I haven't used this before.
