### 2022-10-01 fa54247
minor documentation fixes
