+++ man/right_join.trackr_df.Rd
@@ -18,7 +18,11 @@
 )
 }
 \arguments{
-\item{x, y}{A pair of data frames, data frame extensions (e.g. a tibble), or
+\item{x}{A pair of data frames, data frame extensions (e.g. a tibble), or
+lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
+more details.}
+
+\item{y}{A pair of data frames, data frame extensions (e.g. a tibble), or
 lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
 more details.}
 
