+++ man/p_inner_join.Rd
@@ -17,7 +17,11 @@ p_inner_join(
 )
 }
 \arguments{
-\item{x, y}{A pair of data frames, data frame extensions (e.g. a tibble), or
+\item{x}{A pair of data frames, data frame extensions (e.g. a tibble), or
+lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
+more details.}
+
+\item{y}{A pair of data frames, data frame extensions (e.g. a tibble), or
 lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
 more details.}
 
