+++ man/p_relocate.Rd
@@ -21,7 +21,11 @@ p_relocate(
 
 \item{...}{a set of dplyr summary expressions.}
 
-\item{.before, .after}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Destination of
+\item{.before}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Destination of
+columns selected by \code{...}. Supplying neither will move columns to the
+left-hand side; specifying both is an error.}
+
+\item{.after}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Destination of
 columns selected by \code{...}. Supplying neither will move columns to the
 left-hand side; specifying both is an error.}
 
