### 2022-10-02 4d5ccae
minor documentation fixes
