+++ README.md
@@ -138,7 +138,7 @@ or clinical trials often a requirement.
 dataset %>% flowchart()
 ```
 
-![](data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiIHN0YW5kYWxvbmU9Im5vIj8+CjwhRE9DVFlQRSBzdmcgUFVCTElDICItLy9XM0MvL0RURCBTVkcgMS4xLy9FTiIKICJodHRwOi8vd3d3LnczLm9yZy9HcmFwaGljcy9TVkcvMS4xL0RURC9zdmcxMS5kdGQiPgo8IS0tIEdlbmVyYXRlZCBieSBncmFwaHZpeiB2ZXJzaW9uIDIuNDAuMSAoMjAxNjEyMjUuMDMwNCkKIC0tPgo8IS0tIFRpdGxlOiAlMCBQYWdlczogMSAtLT4KPHN2ZyB3aWR0aD0iODA5cHQiIGhlaWdodD0iMzUwcHQiCiB2aWV3Qm94PSIwLjAwIDAuMDAgODA5LjQ3IDM1MC4wMCIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiB4bWxuczp4bGluaz0iaHR0cDovL3d3dy53My5vcmcvMTk5OS94bGluayI+CjxnIGlkPSJncmFwaDAiIGNsYXNzPSJncmFwaCIgdHJhbnNmb3JtPSJzY2FsZSgxIDEpIHJvdGF0ZSgwKSB0cmFuc2xhdGUoNCAzNDYpIj4KPHRpdGxlPiUwPC90aXRsZT4KPHBvbHlnb24gZmlsbD0iI2ZmZmZmZiIgc3Ryb2tlPSJ0cmFuc3BhcmVudCIgcG9pbnRzPSItNCw0IC00LC0zNDYgODA1LjQ3MTIsLTM0NiA4MDUuNDcxMiw0IC00LDQiLz4KPCEtLSAxNiYjNDU7Jmd0OzE3IC0tPgo8ZyBpZD0iZWRnZTEiIGNsYXNzPSJlZGdlIj4KPHRpdGxlPjE2OnMmIzQ1OyZndDsxNzwvdGl0bGU+CjxwYXRoIGZpbGw9Im5vbmUiIHN0cm9rZT0iIzAwMDAwMCIgZD0iTTMzMi45MDYsLTQxLjRDMzMyLjkwNiwtNDEuNCAzMzIuOTA2LC0yOC40OTg3IDMzMi45MDYsLTI4LjQ5ODciLz4KPHBvbHlnb24gZmlsbD0iIzAwMDAwMCIgc3Ryb2tlPSIjMDAwMDAwIiBwb2ludHM9IjMzNC42NTYxLC0yOC40OTg3IDMzMi45MDYsLTIzLjQ5ODcgMzMxLjE1NjEsLTI4LjQ5ODcgMzM0LjY1NjEsLTI4LjQ5ODciLz4KPC9nPgo8IS0tIDEzJiM0NTsmZ3Q7MTYgLS0+CjxnIGlkPSJlZGdlNCIgY2xhc3M9ImVkZ2UiPgo8dGl0bGU+MTM6cyYjNDU7Jmd0OzE2PC90aXRsZT4KPHBhdGggZmlsbD0ibm9uZSIgc3Ryb2tlPSIjMDAwMDA
[…]
