+++ .github/workflows/R-CMD-check.yaml
@@ -52,7 +52,7 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: 3
-          extra-packages: any::rcmdcheck, any::XML
+          extra-packages: any::rcmdcheck, XML
           # work around for bug in pak
 
       - uses: r-lib/actions/check-r-package@v2
