### 2022-10-04 bb97d5d
refactoring for JOSS review 1
