+++ vignettes/joss/paper.md
@@ -16,34 +16,113 @@ affiliations:
    index: 1
  - name: College of Engineering, Mathematics and Physical Sciences, University of Exeter, Devon, UK
    index: 2
-date: 28 June 2022
+date: 4 October 2022
 bibliography: ../dtrackr.bib
 ---
 
 # Summary
 
-An accurate statement of the provenance of data is essential in bio-medical research. Powerful data manipulation tools available in the `tidyverse` R package ecosystem [@wickhamWelcomeTidyverse2019a] provide the infrastructure to assemble, clean and filter data prior to statistical analysis. Manual documentation of the steps taken in the data pipeline and the provenance of data is a cumbersome and error prone task which may restrict reproducibility. `dtrackr` is a thin wrapper around the standard `tidyverse` data manipulation tools that automatically tracks the processing steps applied to a data set, prior to statistical analysis. It allows early detection and reporting of data quality problems, and automatically documents a set of data transformations as a flowchart in a format suitable for scientific publication, including, but not limited to CONSORT diagrams [@schulzCONSORT2010Statement2010].
+An accurate statement of the provenance of data is essential in bio-medical
+research. Powerful data manipulation tools available in the `tidyverse` R
+package ecosystem [@wickhamWelcomeTidyverse2019a] provide the infrastructure to
+assemble, clean and filter data prior to statistical analysis. Manual
+documentation of the steps taken in the data pipeline and the provenance of data
+is a cumbersome and error prone task which may restrict reproducibility.
+`dtrackr` is a thin wrapper around a subset of the standard `tidyverse` data manipulation
+tools that allows automatic tracking the processing steps applied to a data set,
+prior to statistical analysis. It allows early detection and reporting of data
+quality problems, and automatically documents a set of data transformations as a
+flowchart in a format suitable for scientific publication, including, but not
+limited to CONSORT diagrams [@schulzCONSORT2010Statement2010].
+
+`dtrackr` if first and foremost a utility to accelerate and improve research by
+facilitating documentation, thus supporting extraction of knowledge from data sets,
+and the execution of research by helping identify data quality issues. The
+general capability however fits into a broader context of other provenance or
+data pipeline research. This includes initiatives such as `C2Metadata` [@alterCapturingDataProvenance2021;
+@pimentelSurveyCollectingManaging2019], which focus on a language independent representation of a data
+pipeline, and R packages such as 
+`targets` [@landauTargetsPackageDynamic2021] which focus on documenting
+pipeline code, and managing the execution of a pipeline. `dtrackr` takes a more
+data oriented approach, which could be complementary, in which we remain
+agnostic to the design of a data pipeline. We capture the transformations
+applied to data alongside the data itself, along with documentation of the data
+state as it was transformed. The approach of `dtrackr` is analogous to a `git`
+commit history for dataframes, and there is potential synergy with emerging
+versioned databases such as `dolt` [@DoltGitData2022; @rossDoltrClientDolt2022].
 
 # Statement of need
 
-The collection of experimental or observational data for research is often an iterative endeavour, involving curation of complex data sets designed for multiple goals. Systematic data quality checking for such sets is a major challenge, particularly when they are assembled to identify emerging or rapidly evolving issues. Feedback from early data analysis can identify specific data quality issues, resolution of which can considerably improve data for the task at hand. However this requires a clear understanding of why and when individual data items are excluded, which is potentially tedious and may be seen as lower priority compared to statistical analysis. 
+The collection of experimental or observational data for research is often an
+iterative endeavour, involving curation of complex data sets designed for
+multiple goals. Systematic data quality checking for such sets is a major
[…]
