+++ README.Rmd
@@ -23,6 +23,8 @@ here::i_am("README.Rmd")
 [![R-CMD-check](https://github.com/terminological/dtrackr/workflows/R-CMD-check/badge.svg)](https://github.com/terminological/dtrackr/actions)
 [![DOI](https://zenodo.org/badge/335974323.svg)](https://zenodo.org/badge/latestdoi/335974323)
 [![dtrackr status badge](https://terminological.r-universe.dev/badges/dtrackr)](https://terminological.r-universe.dev)
+[![metacran downloads](https://cranlogs.r-pkg.org/badges/dtrackr)](https://cran.r-project.org/package=dtrackr)
+[![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/dtrackr)](https://cran.r-project.org/package=dtrackr)
 <!-- [![DOI](https://joss.theoj.org/papers/10.21105/joss.04707/status.svg)](https://doi.org/) -->
 <!-- badges: end -->
 
@@ -30,7 +32,7 @@ here::i_am("README.Rmd")
 
 Accurate documentation of a data pipeline is a first step to reproducibility,
 and a flow chart describing the steps taken to prepare data is a useful part of 
-this documentation. In analyses that relies on data that is frequently updated,
+this documentation. In analyses that rely on data that is frequently updated,
 documenting a data flow by copying and pasting row counts into flowcharts in
 PowerPoint becomes quickly tedious. With interactive data analysis, and
 particularly using RMarkdown, code execution sometimes happens in a non-linear
@@ -47,7 +49,7 @@ also a visual check of the actual data processing.
 
 ## Installation
 
-In general use `dtrackr` is expected to be installed alongside the `idyverse`
+In general use `dtrackr` is expected to be installed alongside the `tidyverse`
 set of packages. It is recommended to install `tidyverse` first.
 
 Binary packages of `dtrackr` are available on CRAN and r-universe for `macOS` 
@@ -164,7 +166,8 @@ of `pdf`, `png`, `svg` or `ps`), ready for submission to Nature.
 This is a trivial example, but the more complex the pipeline, the bigger 
 benefit you will get. 
 
-Check out the [main documentation for detailed examples](https://terminological.github.io/dtrackr/)
+Check out the [main documentation for more details](https://terminological.github.io/dtrackr),
+and in particular the [getting started vignette](https://terminological.github.io/dtrackr/articles/dtrackr.html).
 
 ## Testing and integration
 
