### 2022-10-04 terminological/dtrackr#10 @robchallen
Hi. I've worked through all your helpful comments. In the end I made quite a few changes which I will describe in the linked issues, and this has been very worthwhile. I've bundled all the changes into a new release (0.2.4.9000) in a new branch which can be found here:

https://github.com/terminological/dtrackr/tree/joss-fixes-0.2.4.9000

On this specific issue I've [added a LICENCE.md file.](https://github.com/terminological/dtrackr/compare/0.2.4...0.2.4.9000#diff-4673a3aba01813b595de187a7a6e9e63a3491d55821606fecd9f13a10c188a1d)

 This now show up on the home page of GitHub also.

 (p.s. I'll let you close the issues once reviewed)
