### 2022-10-04 terminological/dtrackr#11 @robchallen
I've left the location of the contributing guideline as it is in .github/CONTRIBUTING.md, as in that location it is automatically linked to from the new issues page on GitHub.

To address your other comment about the style of the documentation i've:

- [updated the contents of the CONTRIBUTING.md](https://terminological.github.io/dtrackr/CONTRIBUTING.html)
- [rewritten much of the README.md](https://terminological.github.io/dtrackr/index.html)

Please take a look and see what you think.
