### 2022-10-04 terminological/dtrackr#12 @robchallen
Updated branch: [joss-fixes-0.2.4.9000](https://github.com/terminological/dtrackr/tree/joss-fixes-0.2.4.9000) (and main)

- [rewritten much of the README.md](https://terminological.github.io/dtrackr/index.html) to include a much expanded section on system dependencies. Due to a dependency problem I switched out `staplr` for `pdftools` which has changed the system requirements.
