### 2022-10-04 terminological/dtrackr#13 @robchallen
Updated branch: [joss-fixes-0.2.4.9000](https://github.com/terminological/dtrackr/tree/joss-fixes-0.2.4.9000) (and main)

As I implemented a new set of functionality for  #14 I also extended the functional testing (in the form of additional examples). This has pushed coverage up a bit:

covr::package_coverage(type=“all”): dtrackr Coverage: 83.60%, R/dot.R: 78.12%, R/dtrackr.R: 84.08%
