### 2022-10-04 terminological/dtrackr#14 @robchallen
So I should have addressed this issue first.

There were some gaps in functionality support. These were partly pragmatic and partly due to me not using those functions very much. Part of the implementation though is achieved by sub-classing `data.frame` (as `trackr_df`) and implementing the wrapper functions as S3 methods for the subclass. This means that mixing the native `dplyr` and `dtrackr` functions should just work and can be switched back and forth, and in theory implementing all the `dplyr` methods is not strictly required.

Anyway in the light of this comment I implemented the missing functions in `dplyr`. This triggered me to also address some of the testing issues and improve test coverage (see #13). I have now included `dtrackr` functions for `slice*`, `bind_cols`, `union_all`, `nest_join`.

For documentation of the supported reference I have updated the [function reference page](https://terminological.github.io/dtrackr/reference/index.html) to make it more logically organised and providing some detail about the rationale for the scope. This includes an explanation about what other functions are out of scope, and not yet in scope, compared to `dplyr`.
