### 2022-10-04 terminological/dtrackr#15 @robchallen
Updated branch: [joss-fixes-0.2.4.9000](https://github.com/terminological/dtrackr/tree/joss-fixes-0.2.4.9000) (and main)

I've changed the wording to be more explicit that `dtrackr` is only wrapping a subset of `tidyverse` functions (in `dplyr` and `tidyr`). N.b. see also comments in issue #14 about the fact that it should be possible to interchange between `dplyr`, `dtrackr` and `tidyr` and other `tidyverse` functions within a pipeline.

To address your query about state of the field, I've added a new paragraph into the summary section to describe how I see `dtrackr` fits into the existing landscape, I do see `dtrackr` primarily as a useful tool for researchers, with a somewhat novel take on provenance, rather than a completely new concept, but as a useful tool for research I think it does tick the box. I think future iterations could contribute more to the broader field, particularly if it can be combined with a versioned database as I describe. It was only possible to include very high level summary here as I obviously can't cover off the whole field in the relatively short format of the JOSS article.
