### 2022-10-04 #4707 @robchallen
I've been working through @craig-willis comments, and made quite a few updates. These are in a new branch [joss-fixes-0.2.4.9000](https://github.com/terminological/dtrackr/tree/joss-fixes-0.2.4.9000). Would it be better to do additional review there?
