### 2022-10-05 #4707 @debruine
I’m sorry I got swamped with work and am running late. I’ll start my review in this branch and aim to complete by Monday.

> On 5 Oct 2022, at 00:47, Rob Challen ***@***.***> wrote:
>
> ﻿
> I've been working through @craig-willis comments, and made quite a few updates. These are in a new branch joss-fixes-0.2.4.9000. Would it be better to do additional review there?
>
