### 2022-10-05 #4707 @ajstewartlang
> I've been working through @craig-willis comments, and made quite a few updates. These are in a new branch [joss-fixes-0.2.4.9000](https://github.com/terminological/dtrackr/tree/joss-fixes-0.2.4.9000). Would it be better to do additional review there?

Sure - I'll set that as the review branch now.
