### 2022-10-05 terminological/dtrackr#10 @craig-willis
License looks good.
