### 2022-10-05 terminological/dtrackr#11 @craig-willis
Looks good. Just a note: it looks like a `CODE_OF_CONDUCT.md` in the repo root will display a link under the About section (e.g., https://github.com/github/docs). I wish they did this with with `CONTRIBUTING.md` as well.
