### 2022-10-07 terminological/dtrackr#12 @craig-willis
I ran into a very minor error during the ```pkgdown::build_site()```  noted below, but otherwise everything looks great.

For reference, here are my steps to confirm dtrackr installation on Ubuntu 20.04 via Docker. I tried to clarify which dependencies belong to each step, if useful.

1. Install r-base
2. Install tidyverse
3. Install dtrackr via CRAN
4. Install dtrackr via devtools
5. Install dtrackr from source
6. Run tests/coverage

**1. Install r-base**
```
docker run  -it ubuntu:focal bash

apt update

# Required to install r-base
apt install -y gnupg software-properties-common

# Install r-base
apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 51716619E084DAB9
add-apt-repository 'deb https://cloud.r-project.org/bin/linux/ubuntu focal-cran40/'

apt update
apt install -y r-base

echo "TZ=$( cat /etc/timezone )" >> /etc/R/Renviron.site
```

**2.  Install tidyverse**

```
apt install libcurl4-openssl-dev libssl-dev libxml2-dev

R
> install.packages('tidyverse')
```

**3. Install dtrackr from CRAN**

```
apt install librsvg2-dev libnode-dev

R
> install.packages('dtrackr')
> <Run basic usage example>
> remove.packages('dtrackr')
```

**4. Install dtracker via devtools**

```
apt install libharfbuzz-dev  libfribidi-dev libfreetype6-dev libtiff-dev libpoppler-cpp-dev

R
> install.packages("devtools")
> devtools::install_github("terminological/dtrackr")
> <Run basic usage example>
> remove.packages('dtrackr')
```

**5. Install dtrackr from source**

```
apt-get install git

# Clone repo
git clone https://github.com/terminological/dtrackr
cd dtrackr

R
> install.packages(".", repos=NULL, type="source")
```

**6.  Run tests**
```
> install.packages('pak')
> devtools::load_all()
> pak::local_system_requirements("ubuntu","20.04")
> install.packages(c("here","tidyverse","devtools","testthat","pkgdown"))
> devtools::run_examples()
> devtools::test()
...
== Results ======================================================================
Duration: 174.1 s

[ FAIL 0 | WARN 0 | SKIP 0 | PASS 144 ]
```

**Build site**
```
apt install make pandoc imagemagick libmagick++-dev \
 gsfonts default-jdk  zlib1g-dev libfontconfig1-dev \
 libjpeg-dev libpng-dev libgit2-dev

> pkgdown::build_site()
...
Error in render_rmarkdown(pkg, input = input, output = output_file, output_format = format,  :
  Failed to render RMarkdown
Caused by error:
! error in callr subprocess
Caused by error:
! pandoc document conversion failed with error 83
In addition: Warning message:
Missing images in 'README.md': 'reference/figures/README-flowchart.png?raw=true'
i pkgdown can only use images in 'man/figures' and 'vignettes'

```

**Install covr dependencies**
```
apt install pandoc-citeproc

> install.packages('covr')
> covr::package_coverage(type="all", quiet = FALSE)

...
dtrackr Coverage: 83.60%
R/dot.R: 78.12%
R/dtrackr.R: 84.08%
```
