### 2022-10-07 terminological/dtrackr#13 @craig-willis
Everything works for me as described above and in https://github.com/terminological/dtrackr/issues/12#issuecomment-1271876992. Not a big deal, but you could easily add the covr command to the README, since it works now with a single additional dependency.
