### 2022-10-10 terminological/dtrackr#15 @craig-willis
The emphasis on dtrackr as a utility helps me better understand it's positioning. This and simply signaling awareness of other provenance-related work addresses my main concern.

A few comments:
*  I noticed a possible typo in the first sentence ("dtrackr if first and foremost" -> should be "dtrackr is..." ?)
* Pimentel et al. is not related to C2Metadata. As a broad survey, I guess it's an example of "other provenance research". My main point in sharing the reference was to highlight that there's related work on computational provenance tools -- and even some existing tools that work with R (e.g., RDataTracker, YesWorkflow, recordr) -- that could be used to help frame dtrackr. In section 2.1, the authors discuss a few classifications:  prospective (program/experiment structure) v. retrospective (what actually happened) provenance. In 2.1.4 they discuss execution provenance approaches including passive monitoring, overriding, and instrumentation. In this light, I'd classify dtrackr is a retrospective provenance tool that relies on overriding -- although this probably isn't important for the JOSS paper. I do think the Pimentel reference is useful if only to signal awareness of related work in computational provenance research, but not in the context of C2Metadata.
* C2Metadata is developing a language-independent representation of data transformations via SDTL, but relies on static parsing of scripts (prospective provenance), including R. It doesn't document a "pipeline" in the same sense as dtrackr.  In hindsight, since it's a small/nascent project, and may not something to highlight in the JOSS paper unless it's of interest to you.

A little wordy, but let me know if you have any questions.
