### 2022-10-10 terminological/dtrackr#15 @robchallen
Thanks again,
I'm waiting for the second reviewer's comments. Once I get them I'll take
another stab at this, in light of your additional suggestions. I'm trying
not to go to far down the rabbit hole as I suspect it's pretty deep.
