### 2022-10-16 terminological/dtrackr#18 @debruine
JOSS review - function documentation
Review issue: openjournals/joss-reviews#4707
Branch reviewed: main

All function documentation I've seen assumes the main package is already loaded. It is unusual and unnecessary to begin all examples in the documentation with:

```
library(dplyr)
library(dtrackr)
```

Loading dplyr seems seldom necessary, and would probably be clearer if you use `dplyr::func()` to demark any functions that don't come from dtrackr in your examples. I tried all of the examples in the "Controlling dtrackr" section and none required dplyr (the %>% pipe is loaded with dtrackr).
