### 2022-10-16 terminological/dtrackr#19 @debruine
JOSS review - Automated tests
Review issue: openjournals/joss-reviews#4707
Branch reviewed: main

I'm not sure the testing and integration section is necessary on the [main documentation page](https://terminological.github.io/dtrackr/). This is likely to confuse most R users, and users with the interest and experience necessary to do this will already know how.

I know a section of the JOSS review asks "Are there automated tests or manual steps described so that the functionality of the software can be verified?", but this disrupts the expected structure of R package documentation and other recent JOSS R packages that use pkgdown (e.g., https://joss.theoj.org/papers/10.21105/joss.04684) don't include this.
