### 2022-10-16 terminological/dtrackr#20 @debruine
JOSS review - Documentation typo
Review issue: openjournals/joss-reviews#4707
Branch reviewed: main

The hex icon on the [main page](https://terminological.github.io/dtrackr/index.html) links to [dplyr](https://dplyr.tidyverse.org/), which seems like a typo.
