### 2022-10-16 terminological/dtrackr#21 @debruine
JOSS Review - Dataset Documentation
Review issue: openjournals/joss-reviews#4707
Branch reviewed: main

The [ILPD dataset](https://terminological.github.io/dtrackr/reference/ILPD.html) needs clearer explanations of each column. I know this is just example data and none of this really matters for your purpose of having demo data, but I'm on the rampage against poorly documented data in R packages.

For example, you could give the category labels for Gender and specify if Age is in years and give the range. I checked your source for the dataset and it specifies an interesting caveat that you may want to mention directly:

> Any patient whose age exceeded 89 is listed as being of age "90".

I was going to ask you to give the units for the blood measurements, but I tried looking these up myself and they are maddeningly hard to find! Maybe just give the range for reference?
