### 2022-10-16 terminological/dtrackr#22 @debruine
JOSS Review - Function Documentation for landscape
Review issue: openjournals/joss-reviews#4707
Branch reviewed: main

The documentation for [`landscape()`](https://terminological.github.io/dtrackr/reference/landscape.html) is quite sparse. It could use an example. My first attempt to guess failed (I tried `landscape(std_size$A4$height)`) and that was only after I'd read the documentation for `std_size`. You might link to `std_size`, as it's not immediately obvious that this is an object built into dtrackr.

It's also not obvious when or where you'd use this function. It doesn't take any dtrackr objects as an argument, so I first assumed it was an option that needed to be set before running `save_dot()`, but that didn't work. I noticed it outputs a size list, so then I tried the code below, but that doesn't seem to produce any difference in orientation:

```
x <- save_dot("digraph {A->B}", tempfile(),
              size = landscape(std_size$half),
              formats = "png")
browseURL(x$paths$png)
```
