### 2022-10-16 #4707 @debruine
I'm finding this review process really interesting but a little overwhelming (I'm never sure I've read all the relevant instructions or are doing everything correctly). I'm nearly done checking the documentation for each function and hope to finish by Tuesday.

I just wanted to say that I love this package and have immediately started using it in my own data wrangling pipeline. It's very intuitive for a tidyverse user. I wish this were around 5 years ago when I was printing little tables after each wrangling step in a super-complex data join and filter.
