### 2022-10-17 #4707 @robchallen
> I just wanted to say that I love this package and have immediately started using it in my own data wrangling pipeline. It's very intuitive for a tidyverse user. I wish this were around 5 years ago when I was printing little tables after each wrangling step in a super-complex data join and filter.

Thanks! This is surely the most positive peer review comment I've ever had, and made my day! I'm really pleased to hear it is useful.
