+++ inst/examples/set-operation-examples.R
@@ -1,5 +1,4 @@
 library(dplyr)
-library(tidyr)
 library(dtrackr)
 
 # Set operations
