+++ NAMESPACE
@@ -113,7 +113,6 @@ export(tagged)
 export(track)
 export(untrack)
 importFrom(dplyr,add_count)
-importFrom(dplyr,add_tally)
 importFrom(dplyr,anti_join)
 importFrom(dplyr,arrange)
 importFrom(dplyr,distinct)
