### 2022-11-04 terminological/dtrackr#13 @robchallen
Just to update on this I found out the right magic incantation to get the `covr` output hooked up to github actions and the whole thing displayed as a badge on the README. simple when you know how :-)
