### 2022-11-04 terminological/dtrackr#14 @robchallen
I don't think there is anything outstanding on this issue so closing for now. Please reopen if needed.
