### 2022-11-04 terminological/dtrackr#17 @robchallen
Sorry for delay in this (half term).

You are of course entirely correct and it was not the best piece of documentation I've written. I've fully rewritten the exclude_all documentation both in the [function reference](https://terminological.github.io/dtrackr/reference/exclude_all.html) and in the [supporting vignette](https://terminological.github.io/dtrackr/articles/dtrackr.html#filtering-exclusions-and-inclusions) to make this clearer including an example as you suggested.

There was a similar problem in the [include_any documentation](https://terminological.github.io/dtrackr/reference/include_any.html) which I have rectified also and included similar example.

In both cases I have drawn out the parallels between these functions and the vanilla `dplyr::filter` functions.
