### 2022-11-04 terminological/dtrackr#19 @robchallen
Agreed.

I have moved the details about the testing set up to the contributors guide as there was a sensible place for it there:
https://terminological.github.io/dtrackr/CONTRIBUTING.html.

The README is now a bit more succinct (https://terminological.github.io/dtrackr/index.html).

@craig-willis FYI as this was something we discussed previously.
