### 2022-11-04 terminological/dtrackr#21 @robchallen
As you say the ILPD dataset was just for example.

I came across the `survival` packages `gcd` dataset through which is a better example of the kind of thing I wanted to demonstrate so switched the examples to using `survival::gcd`.

This means I don't need to bundle the ILPD dataset, and can hence dodge the responsibility of documenting it :-)

The main place this change is seen is in [the vignette](https://terminological.github.io/dtrackr/articles/consort-example.html) which now gives an example using `cgd`.

Since this is now gone I'm going to close this issue. Let me know if any concerns.
