### 2022-11-04 terminological/dtrackr#22 @robchallen
Apologies. The landscape function did do something useful in an earlier iteration of `dtrackr`, and resulted in pdf and png formats being rotated, but the functionality got taken out as was not really core and the dependencies it introduced were complex.

To address this specific issue I've taken out the `landscape()` function.

What I have done instead is extend the `std_size` definitions to include some landscape style paper dimensions. I've linked to the `std_size` documentation from [flowchart()](https://terminological.github.io/dtrackr/reference/flowchart.html) to make this more explicit, and I've updated the `std_size` documentation to make it clear what it is, what values are available, and what it is for.
