### 2022-11-04 terminological/dtrackr#22 @robchallen
std_size documentation is at https://terminological.github.io/dtrackr/reference/std_size.html

Please close this issue if you are happy with this outcome.
