### 2022-11-04 #4707 @robchallen
Sorry for delays responding to comments. I've fixed everything I can now and thanks to your pointers think the newest version is a big improvement.

All the changes are currently in the main branch under the version 0.2.4.9001. I'll leave the version as is until you've had a chance to look, and if/when you are happy with it I will bump it to 0.2.5 and release it to r-universe and submit update to CRAN.

Thanks again for putting in the effort to review this. I really appreciate it.
