### 2022-11-05 terminological/dtrackr#19 @craig-willis
This all looks good to me.  I was thinking more from the contributor perspective than user.
