### 2022-11-14 terminological/dtrackr#22 @robchallen
@debruine Is there anything else you need me to take a look or are you happy to sign off on this and the [other open issue](#17)
All the best,
