### 2022-11-21 #4707 @robchallen
@debruine, I think you may still need to tick one more box on this page if everything OK from your perspective. N.B. I'm about to submit updates to CRAN.
