### 2022-11-21 #4707 @ajstewartlang
Many thanks for your reviews @debruine and @craig-willis - can I check you're both now happy that all issues have been addressed and for @robchallen to bump the JOSS submission up  to 0.2.5?
