+++ R/dtrackr.R
@@ -394,12 +394,21 @@ plot.trackr_graph = function(x, fill="lightgrey", fontsize="8", colour="black", 
 #' Start tracking the dtrackr history graph
 #'
 #' @param .data a dataframe which may be grouped
-#' @param .messages a character vector of glue specifications. A glue specification can refer to any grouping variables of .data, or any variables defined in the calling environment, the \{.total\} variable which is the count of all rows,
-#' the \{.count\} variable which is the count of rows in the current group and the \{.strata\} which describes the current group. Defaults to the value of `getOption("dtrackr.default_message")`.
-#' @param .headline a glue specification which can refer to grouping variables of .data, or any variables defined in the calling environment, or the \{.total\} variable which is nrow(.data), or \{.strata\} a summary of the current group. Defaults to the value of getOption("dtrackr.default_headline").
-#' @param .tag if you want the summary data from this step in the future then give it a name with .tag.
+#' @param .messages a character vector of glue specifications. A glue
+#'   specification can refer to any grouping variables of .data, or any
+#'   variables defined in the calling environment, the \{.total\} variable which
+#'   is the count of all rows, the \{.count\} variable which is the count of
+#'   rows in the current group and the \{.strata\} which describes the current
+#'   group. Defaults to the value of `getOption("dtrackr.default_message")`.
+#' @param .headline a glue specification which can refer to grouping variables
+#'   of .data, or any variables defined in the calling environment, or the
+#'   \{.total\} variable which is `nrow(.data)`, or \{.strata\} a summary of the
+#'   current group. Defaults to the value of `getOption("dtrackr.default_headline")`.
+#' @param .tag if you want the summary data from this step in the future then
+#'   give it a name with .tag.
 #'
-#' @return the .data dataframe with additional history graph metadata, to allow tracking.
+#' @return the .data dataframe with additional history graph metadata, to allow
+#'   tracking.
 #' @export
 #'
 #' @examples
@@ -489,9 +498,13 @@ p_resume = function(.data) {
 #' Start capturing exclusions on a tracked dataframe.
 #'
 #' @param .data a tracked dataframe
-#' @param .capture Should we capture exclusions (things removed from the data set). This is useful for debugging data issues but comes at a significant cost. Defaults to the value of `getOption("dtrackr.exclusions")` or `FALSE`.
+#' @param .capture Should we capture exclusions (things removed from the data
+#'   set). This is useful for debugging data issues but comes at a significant
+#'   cost. Defaults to the value of `getOption("dtrackr.exclusions")` or
+#'   `FALSE`.
 #'
-#' @return the .data dataframe with the exclusions flag set (or cleared if `.capture=FALSE`).
+#' @return the .data dataframe with the exclusions flag set (or cleared if
+#'   `.capture=FALSE`).
 #' @export
 #'
 #' @examples
@@ -519,11 +532,11 @@ p_capture_exclusions = function(.data, .capture=TRUE) {
 #'
 #' @param .data a dataframe which may be grouped
 #'
-#' @return the history graph. This is a list, of class trackr_graph, containing the following named items:
+#' @return the history graph. This is a list, of class `trackr_graph`, containing the following named items:
 #' * excluded - the data items that have been excluded thus far as a nested dataframe
 #' * tags - a dataframe of tag-value pairs containing the summary of the data at named points in the data flow (see [tagged()])
 #' * nodes - a dataframe of the nodes of the flow chart
-#' * edges - an edgelist (as a dataframe) of the relationships between the nodes in the flow chart
[…]
