+++ man/count_subgroup.Rd
@@ -22,7 +22,7 @@ count_subgroup(
 \item{.subgroup}{a column with a small number of levels (e.g. a factor)}
 
 \item{...}{additional parameters will be passed to \code{factor(subgroup,...)} to
-control levels, ordering, etc.}
+control levels, ordering, etcetera.}
 
 \item{.messages}{a character vector of glue specifications. A glue
 specification can refer to anything from the calling environment and \{.name\}
