### 2022-11-23 #4707 @ajstewartlang
@robchallen can I check that you've fixed that typo? Can you also confirm you've bumped the JOSS version to 0.2.5? I'll then set that as the branch and do a few final checks. Thanks for a great submission!
