### 2022-11-23 #4707 @robchallen
Thanks @ajstewartlang, @craig-willis and @debruine. Really useful to get the comments and feedback and I learnt a lot in the process.

I'm not sure what order to do this all but yes I have fixed the typo(s)  in `release-0.2.5` branch, I have deployed that version to the R-universe and will now submit as an update to CRAN.
