### 2022-11-24 #4707 @ajstewartlang
:wave:  @robchallen

If you could now do the following please, that would be great:

- Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
