### 2022-11-25 #4707 @robchallen
10.5281/zenodo.7352369

https://github.com/terminological/dtrackr/releases/tag/0.2.5-cran-submission

@ajstewartlang Many thanks. I think I have done this bit right - it looks all correct on zenodo.
