### 2022-11-29 #4707 @Kevin-Mattheus-Moerman
@robchallen I am the AEiC for this submission and here to help process this submission for acceptance in JOSS. Below are some minor points that require your attention.

General points:
- [ ] We require that the version tags on your **repository**, the version tag listed for the **JOSS review and paper**, and the version tag listed on **ZENODO** match. Currently for this review had 'v0.2.5' (note the `v`) while on your repository we have both `0.2.5` (note no `v`) and `0.2.5-cran-submission`, and the ZENODO archive shows `0.2.5-cran-submission`.
I have just altered the JOSS review/paper version tag :point_up: to match `0.2.5-cran-submission`. Let me know if that is okay or if you'd like to address our requirement in a different way.

On the paper:
- [ ] Please spell out `UK` as `United Kingdom` in your affiliations.
- [ ] You may use `bio-medical` if you wish, but I would say `biomedical` is more common so would recommend the latter. If you agree please update throughout.

On the license:
- [ ] Your license is unclear at the moment since you have both a `LICENSE` and a `LICENSE.md` file. The former contains only a copyright notice which does not grant open source permission. The second then is an MIT license (with its own copyright header already) which does grant open source permissions. The two together is confusing. GitHub currently also lists the license currently as ` Unknown, MIT licenses found`. Please remove this file: https://github.com/terminological/dtrackr/blob/main/LICENSE and keep only the MIT license file. Next I recommend renaming `LICENSE.md` to `LICENSE` (and have it be a plain text license file). This will also help GitHub detect the license properly and render the clear license badge. If you make this change you may wish to also create a new tagged release and associated ZENODO listing, and thereby address the issue at the top here as well. If you do please let me know.
