### 2022-11-29 #4707 @Kevin-Mattheus-Moerman
>From my the perspective the "0.2.5-cran-submission" release is fine - however the zenodo archive automatically gets updated whenever I make a github release. I do need to fix a URL for CRAN as well as those two tweaks to the paper shall I fix those and create a new release and give you a new zenodo DOI?

That sounds good. Sorry this is an awkward process.

> For the license issue, the problem is that CRAN require that the LICENSE file for MIT licensed content contain only the licensor / year and not the full details which must be supplied separately in the LICENSE.md field. Any change to this will result in the CRAN submission getting rejected. if you take a look at another MIT licensed CRAN package (e.g. https://github.com/tidyverse/dplyr/) you'll see this same pattern and GitHub similarly doesn't pick up the license.
>
> Basically I can't change the way the LICENSE files are without breaking the CRAN submission which is insanely particular about what it will accept, and there is no way to simultaneously make GitHub and CRAN happy. Can you live with them the way they are?

Okay. Yes, just leave things (in terms of the license stuff) the way they are then.
