### 2022-12-05 8094c8d
minor fix urls for cran and spelling for JOSS
