### 2022-12-05 ab76b1e
minor fix urls for cran and spelling for JOSS
