+++ README.Rmd
@@ -168,6 +168,6 @@ of `pdf`, `png`, `svg` or `ps`), ready for submission to Nature.
 This is a trivial example, but the more complex the pipeline, the bigger 
 benefit you will get. 
 
-Check out the [main documentation for more details](https://terminological.github.io/dtrackr),
+Check out the [main documentation for more details](https://terminological.github.io/dtrackr/),
 and in particular the [getting started vignette](https://terminological.github.io/dtrackr/articles/dtrackr.html).
 
