### 2022-12-05 f189741
minor fix urls for cran
