+++ R/dtrackr.R
@@ -127,6 +127,8 @@
     capture = current$capture,
     tags = current$tags
   )
+  #TODO: better way of doing a copy operation of tracker_graph
+  class(new) = c("trackr_graph",class(new))
   new$nodes = current$nodes %>% dplyr::bind_rows(nodes)
   #TODO: investigate .rel here. I think it should be more to do with .asOffshoot than defined by .type
   newEdges = current$head %>%
@@ -248,7 +250,12 @@ print.trackr_graph = function(x, ...) {
   if (is.null(graph$excluded)) {
     excluded = "<not capturing exclusions>"
   } else {
-    excluded = sum(sapply(graph$excluded$.excluded, nrow ), na.rm = TRUE)
+    listOfDf = graph$excluded$.excluded
+    if (length(listOfDf) > 0) {
+      excluded = sum(sapply(listOfDf, nrow ), na.rm = TRUE)
+    } else {
+      excluded = 0
+    }
   }
 
   tmp = c(
@@ -731,8 +738,16 @@ p_clear = function(.data) {
 
 
 #' mtcars %>% p_copy(iris %>% comment("A comment")) %>% history()
-p_copy = function(.data, from) {
-  return(.data %>% p_set(from %>% p_get()))
+p_copy = function(.data, from, preserve_others = FALSE) {
+  tmp = .data %>% p_set(from %>% p_get())
+  if (preserve_others) {
+    old_attr = setdiff(
+      names(attributes(from)),
+      # these structural attributes are from dplyr group
+      c(names(attributes(tmp)),"row.names","names","groups")
+    tmp = tmp %>% magrittr::set_attributes(attributes(from)[old_attr])
+  }
+  return(tmp)
 }
 
 ## User operations ----
@@ -1453,7 +1468,7 @@ p_pivot_wider = function(data, id_cols = NULL, names_from = as.symbol("name"), n
     ...
   )
   # TODO: shold this be a .beforeAfterGroupwiseCount operation as it goes from narrow to long
-  out = out %>% p_copy(.data) %>% .comment(.messages, .headline = .headline, .type="pivot_wider", .tag=.tag)
+  out = out %>% p_copy(.data, preserve_others = TRUE) %>% .comment(.messages, .headline = .headline, .type="pivot_wider", .tag=.tag)
   return(out %>% .retrack())
 }
 
@@ -1504,7 +1519,75 @@ p_pivot_longer = function(data,
     values_transform = values_transform,
     ...
   )
-  out = out %>% p_copy(.data) %>% .comment(.messages, .headline = .headline, .type="pivot_longer", .tag=.tag)
+  out = out %>% p_copy(.data, preserve_others = TRUE) %>% .comment(.messages, .headline = .headline, .type="pivot_longer", .tag=.tag)
+  return(out %>% .retrack())
+}
+
+#' Reshaping data using `tidyr::unnest`
+#'
+#' A drop in replacement for [tidyr::unnest()] which optionally takes a
+#' message and headline to store in the history graph.
+#' @seealso tidyr::unnest()
+#'
+#' @inheritParams tidyr::unnest
+#' @param .messages a set of glue specs. The glue code can use any global
+#'   variable, grouping variable, or \{.strata\}. Defaults to nothing.
+#' @param .headline a headline glue spec. The glue code can use any global
+#'   variable, grouping variable, or \{.strata\}. Defaults to nothing.
+#' @param .tag if you want the summary data from this step in the future then
+#'   give it a name with .tag.
+#'
+#' @return the data dataframe result of the `tidyr::unnest` function but with
[…]
