+++ man/unnest.Rd
@@ -1,0 +1,113 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dtrackr.R
+\name{unnest}
+\alias{unnest}
+\title{Reshaping data using \code{tidyr::unnest}}
+\usage{
+unnest(
+  data,
+  cols,
+  ...,
+  keep_empty = FALSE,
+  ptype = NULL,
+  names_sep = NULL,
+  names_repair = "check_unique",
+  .drop = deprecated(),
+  .id = deprecated(),
+  .sep = deprecated(),
+  .preserve = deprecated(),
+  .messages = "",
+  .headline = "",
+  .tag = NULL
+)
+}
+\arguments{
+\item{data}{A data frame.}
+
+\item{cols}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to unnest.
+
+If you \code{unnest()} multiple columns, parallel entries must be of
+compatible sizes, i.e. they're either equal or length 1 (following the
+standard tidyverse recycling rules).}
+
+\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to nest, specified
+using name-variable pairs of the form \code{new_col = c(col1, col2, col3)}.
+The right hand side can be any valid tidy select expression.
+
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}:
+previously you could write \code{df \%>\% nest(x, y, z)} and \code{df \%>\% unnest(x, y, z)}. Convert to \code{df \%>\% nest(data = c(x, y, z))}.
+and \code{df \%>\% unnest(c(x, y, z))}.
+
+If you previously created new variable in \code{unnest()} you'll now need to
+do it explicitly with \code{mutate()}. Convert \code{df \%>\% unnest(y = fun(x, y, z))}
+to \code{df \%>\% mutate(y = fun(x, y, z)) \%>\% unnest(y)}.}
+
+\item{keep_empty}{By default, you get one row of output for each element
+of the list your unchopping/unnesting. This means that if there's a
+size-0 element (like \code{NULL} or an empty data frame), that entire row
+will be dropped from the output. If you want to preserve all rows,
+use \code{keep_empty = TRUE} to replace size-0 elements with a single row
+of missing values.}
+
+\item{ptype}{Optionally, a named list of column name-prototype pairs to
+coerce \code{cols} to, overriding the default that will be guessed from
+combining the individual values. Alternatively, a single empty ptype
+can be supplied, which will be applied to all \code{cols}.}
+
+\item{names_sep}{If \code{NULL}, the default, the names will be left
+as is. In \code{nest()}, inner names will come from the former outer names;
+in \code{unnest()}, the new outer names will come from the inner names.
+
+If a string, the inner and outer names will be used together. In
+\code{unnest()}, the names of the new outer columns will be formed by pasting
+together the outer and the inner column names, separated by \code{names_sep}. In
+\code{nest()}, the new inner names will have the outer names + \code{names_sep}
+automatically stripped. This makes \code{names_sep} roughly symmetric between
+nesting and unnesting.}
+
+\item{names_repair}{Used to check that output data frame has valid
+names. Must be one of the following options:
+\itemize{
+\item "minimal": no name repair or checks, beyond basic existence,
+\item "unique": make sure names are unique and not empty,
+\item "check_unique": (the default), no name repair, but check they are unique,
[…]
