+++ man/p_nest.Rd
@@ -1,0 +1,65 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dtrackr.R
+\name{p_nest}
+\alias{p_nest}
+\title{Nesting data using \code{tidyr::nest}}
+\usage{
+p_nest(
+  .data,
+  ...,
+  .names_sep = NULL,
+  .key = NULL,
+  .messages = "",
+  .headline = "",
+  .tag = NULL
+)
+}
+\arguments{
+\item{.data}{A data frame.}
+
+\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to nest, specified
+using name-variable pairs of the form \code{new_col = c(col1, col2, col3)}.
+The right hand side can be any valid tidy select expression.
+
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}:
+previously you could write \code{df \%>\% nest(x, y, z)} and \code{df \%>\% unnest(x, y, z)}. Convert to \code{df \%>\% nest(data = c(x, y, z))}.
+and \code{df \%>\% unnest(c(x, y, z))}.
+
+If you previously created new variable in \code{unnest()} you'll now need to
+do it explicitly with \code{mutate()}. Convert \code{df \%>\% unnest(y = fun(x, y, z))}
+to \code{df \%>\% mutate(y = fun(x, y, z)) \%>\% unnest(y)}.}
+
+\item{.names_sep}{If \code{NULL}, the default, the names will be left
+as is. In \code{nest()}, inner names will come from the former outer names;
+in \code{unnest()}, the new outer names will come from the inner names.
+
+If a string, the inner and outer names will be used together. In
+\code{unnest()}, the names of the new outer columns will be formed by pasting
+together the outer and the inner column names, separated by \code{names_sep}. In
+\code{nest()}, the new inner names will have the outer names + \code{names_sep}
+automatically stripped. This makes \code{names_sep} roughly symmetric between
+nesting and unnesting.}
+
+\item{.key}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}:
+No longer needed because of the new \code{new_col = c(col1, col2, col3)} syntax.}
+
+\item{.messages}{a set of glue specs. The glue code can use any global
+variable, grouping variable, or \{.strata\}. Defaults to nothing.}
+
+\item{.headline}{a headline glue spec. The glue code can use any global
+variable, grouping variable, or \{.strata\}. Defaults to nothing.}
+
+\item{.tag}{if you want the summary data from this step in the future then
+give it a name with .tag.}
+}
+\value{
+the result of the \code{tidyr::nest} but with a history graph
+updated.
+}
+\description{
+A drop in replacement for \code{\link[tidyr:nest]{tidyr::nest()}} which optionally takes a
+message and headline to store in the history graph.
+}
+\seealso{
+tidyr::nest()
+}
