+++ NAMESPACE
@@ -49,6 +49,7 @@ export(filter)
 export(flowchart)
 export(history)
 export(include_any)
+export(nest)
 export(p_add_count)
 export(p_add_tally)
 export(p_anti_join)
@@ -76,6 +77,7 @@ export(p_inner_join)
 export(p_intersect)
 export(p_left_join)
 export(p_mutate)
+export(p_nest)
 export(p_nest_join)
 export(p_pause)
 export(p_pivot_longer)
@@ -103,6 +105,7 @@ export(p_transmute)
 export(p_ungroup)
 export(p_union)
 export(p_union_all)
+export(p_unnest)
 export(p_untrack)
 export(pause)
 export(resume)
@@ -111,6 +114,7 @@ export(status)
 export(std_size)
 export(tagged)
 export(track)
+export(unnest)
 export(untrack)
 importFrom(dplyr,add_count)
 importFrom(dplyr,anti_join)
@@ -144,5 +148,7 @@ importFrom(dplyr,ungroup)
 importFrom(dplyr,union)
 importFrom(dplyr,union_all)
 importFrom(magrittr,"%>%")
+importFrom(tidyr,nest)
 importFrom(tidyr,pivot_longer)
 importFrom(tidyr,pivot_wider)
+importFrom(tidyr,unnest)

[…]
