### 2022-12-12 #4707 @robchallen
@Kevin-Mattheus-Moerman Hi sorry for extended delay. I'm waiting on CRAN to approve my change of email address. I decided it was nopt a good idea to try and wait for them so I plan to go ahead and release everything today and update zenodo.

I just wanted to check before I do that about the citation info.

Currently this is pointing at the Zenodo releases and Github repo, but CRAN has had a  few issues with the Zenodo DOIs (another source of delay).

I thought it would make sense to update to the (future) JOSS article. From what I can see this is what I think the citation will be:

```
@Article{,
  title = {dtrackr: An R package for tracking the provenance of data},
  year = {2022},
  note = {R package version 0.2.5},
  url = {https://joss.theoj.org/papers/10.21105/joss.04707},
  doi = {10.21105/joss.04707},
  journal = {Journal of Open Source Software},
  author = {Robert Challen},
}
```

Could you confirm this is OK, and I will go ahead and release it.
Many thanks and sorry again for the extended delay.
