### 2022-12-13 #4707 @Kevin-Mattheus-Moerman
@robchallen yes this :point_down: looks accurate:
```
@Article{,
  title = {dtrackr: An R package for tracking the provenance of data},
  year = {2022},
  note = {R package version 0.2.5},
  url = {https://joss.theoj.org/papers/10.21105/joss.04707},
  doi = {10.21105/joss.04707},
  journal = {Journal of Open Source Software},
  author = {Robert Challen},
}`
```
