### 2022-12-13 3d8ec56
minor documentation / paper update
