+++ DESCRIPTION
@@ -46,7 +46,7 @@ VignetteBuilder:
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.2
 Depends: 
     R (>= 2.10)
 URL: https://terminological.github.io/dtrackr/index.html,
