### 2022-12-13 #4707 @robchallen
@Kevin-Mattheus-Moerman
Thank you for confirming the CITATION info. I have:

- [ ] made the change to the CITATION file,
- [ ] updated UK to United Kingdom
- [ ] updated "bio-medical" to "biomedical"

There are a variety of other typo fixes in the package documentation triggered by the CRAN submission.

I've released it and checked the updated zenodo metadata matches the JOSS metadata:

New github tag: [0.2.5-joss](https://github.com/terminological/dtrackr/releases/tag/0.2.5-joss)
New Zenodo record: [DOI: 10.5281/zenodo.7433514](https://zenodo.org/record/7433514#.Y5iaf9LP3mE)

Please let me know if there is anything else I can do, and thanks again.
