### 2022-12-13 #4707 @Kevin-Mattheus-Moerman
@robchallen nearly there. Can you fix the license on the ZENODO archive, it says `Other (open)` at the moment and should be the same as your actual software license.
