### 2022-12-13 #4707 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
